backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6
PyYAML==6.0.1
aiohttp==3.9.1
mlx-lm==0.28.3
